Fix: Don't reject an already present user (#28)
The socket JS client leaving the channel is asynchronous. This makes it impossible to simply disconnect and connect again from React. This situation, without this fix, causes the page to not work about 60% of the times, because the channel would reject a connection from an already present username. With this fix, it accepts the connection.